feat(home): init getalllayout
