--- src/components/GetAllCard/styles.ts
@@ -1,7 +1,0 @@
-import { css } from "@emotion/react";
-
-const GetAllCard = css`
-  margin: 100px 0;
-`;
-
-export default GetAllCard;
